Convert to Open Document Spreadsheet, for compatibility. ========================================================= Include the XLS commit message, so it's attached to this file as well:
Didn't prove it, but convinced myself that the square laying flat on the triangle's base will give the biggest square inside a triangular hole, and it requires square-side = q <= 0.464 to just barely fit, or fall thru.  Thus, if the square's edge is at least half of the triangular hole base... and the triangular hole base is half of the tetrahedron's triangular base, it would take a cube with edge < 14.848 to fall thru, but anything under 16 is pretty unstable, even if it doesn't fall thru.  

Thus, I could shrink my Menger Sponge to E=1 L=3 => Q = 1 * 3^3 = 27 without difficulty... but the question becomes, do I want to?  Probably not.  Still, I'll archive this math.