add a couple comments to the app.pl
Ran NYTProf on it -- took a long time, but finished just before leaving for the day.  Add the report (but not the 11G database), will study it later...

I thought I had deleted MovieID>100 from the database, but apparently the Firefox SQLite plugin I was using doesn't save the database.  That's why profiler took so long.