add CONFIG_REQUIRES prereqs that were missing
this won't help in GitHub Actions environment, because the META files are generated by MakeMaker when Makefile.PL is run... but Makefile.PL won't run without them (chicken and egg problem).  But I figured that since the helper module used those, I should include it for correctness.

Leaving AppVeyor enabled this time, because with the Action failing intermittently, I want to make sure my change to the helper hasn't broken the normal process.